permisos para Admin para inventario y restricciones a secretaria y odontologo

+++ routes/web.php
@@ -260,12 +260,12 @@ Route::get('pantallainicio/odontologo/{id}','EspecialidadOdontologosController@E
 
 
 //**************Rutas de  inventarios******************/
-Route::get('inventario','InventarioController@vistaprincipal');
-Route::delete('inventario/{id}/borrar','InventarioController@destroy') ->name('inventario.borrar')->where('id','[0-9]+');
-Route::post('inventarioNuevo','InventarioController@guardar');
+Route::get('inventario','InventarioController@vistaprincipal')->middleware('role:admin,secretaria');
+Route::delete('inventario/{id}/borrar','InventarioController@destroy') ->name('inventario.borrar')->where('id','[0-9]+')->middleware('role:secretaria,admin');
+Route::post('inventarioNuevo','InventarioController@guardar')->middleware('role:admin,secretaria');
 Route::get('inventarionuevo','InventarioController@nuevo');
-Route::get('inventario/{id}/editar','InventarioController@editar') ->name('inventario.editar') -> where('id' ,'[0-9]+');
-Route::put('inventario/{id}/editar','InventarioController@update')->name('inventario.update') -> where('id' ,'[0-9]+');
+Route::get('inventario/{id}/editar','InventarioController@editar') ->name('inventario.editar') -> where('id' ,'[0-9]+')->middleware('role:secretaria,admin');
+Route::put('inventario/{id}/editar','InventarioController@update')->name('inventario.update') -> where('id' ,'[0-9]+')->middleware('role:secretaria,admin');
 
 //**************Rutas de  medios de pago******************/
 Route::get('/pantallainicio/mediopago','MediodepagoController@vistaprincipal')->middleware('role:admin,secretaria');
